+++ Constraid/ManageIntrinsicSizeRelations.swift
@@ -9,7 +9,7 @@ extension Constraid.View {
         Sets height content hugging and compression resistance priority to 1000
         (required)
     */
-    open func keepIntrinsicHeight(priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+    open func keepIntrinsicHeight(priority: Constraid.LayoutPriority = Constraid.LayoutPriorityRequired) {
         self.setContentCompressionResistancePriority(priority, for: .vertical)
         self.setContentHuggingPriority(priority, for: .vertical)
     }
@@ -18,7 +18,7 @@ extension Constraid.View {
         Sets width content hugging and compression resistance priority to 1000
         (required)
     */
-    open func keepIntrinsicWidth(priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+    open func keepIntrinsicWidth(priority: Constraid.LayoutPriority = Constraid.LayoutPriorityRequired) {
         self.setContentCompressionResistancePriority(priority, for: .horizontal)
         self.setContentHuggingPriority(priority, for: .horizontal)
     }
@@ -27,7 +27,7 @@ extension Constraid.View {
         Sets height & width content hugging and compression resistance priority to 1000
         (required)
     */
-    open func keepIntrinsicSize(priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+    open func keepIntrinsicSize(priority: Constraid.LayoutPriority = Constraid.LayoutPriorityRequired) {
         keepIntrinsicHeight(priority: priority)
         keepIntrinsicWidth(priority: priority)
     }