+++ ConstraidTests/ManageIntrinsicSizeRelationsTests.swift
@@ -5,7 +5,7 @@ class ManageIntrinsicSizeRelationsTests: XCTestCase {
     func testKeepIntrinsicHeight() {
         let viewOne = UIView()
 
-        viewOne.keepIntrinsicHeight(priority: ConstraidLayoutPriority(rawValue: 284))
+        viewOne.keepIntrinsicHeight(priority: Constraid.LayoutPriority(rawValue: 284))
 
         XCTAssertEqual(viewOne.contentCompressionResistancePriority(for: .vertical), UILayoutPriority(rawValue: UILayoutPriority.RawValue(284)))
         XCTAssertEqual(viewOne.contentHuggingPriority(for: .vertical), UILayoutPriority(rawValue: UILayoutPriority.RawValue(284)))
@@ -14,7 +14,7 @@ class ManageIntrinsicSizeRelationsTests: XCTestCase {
     func testKeepIntrinsicWidth() {
         let viewOne = UIView()
 
-        viewOne.keepIntrinsicWidth(priority: ConstraidLayoutPriority(rawValue: 284))
+        viewOne.keepIntrinsicWidth(priority: Constraid.LayoutPriority(rawValue: 284))
 
         XCTAssertEqual(viewOne.contentCompressionResistancePriority(for: .horizontal), UILayoutPriority(rawValue: UILayoutPriority.RawValue(284)))
         XCTAssertEqual(viewOne.contentHuggingPriority(for: .horizontal), UILayoutPriority(rawValue: UILayoutPriority.RawValue(284)))
@@ -23,7 +23,7 @@ class ManageIntrinsicSizeRelationsTests: XCTestCase {
     func testKeepIntrinsicSize() {
         let viewOne = UIView()
 
-        viewOne.keepIntrinsicSize(priority: ConstraidLayoutPriority(rawValue: 284))
+        viewOne.keepIntrinsicSize(priority: Constraid.LayoutPriority(rawValue: 284))
 
         XCTAssertEqual(viewOne.contentCompressionResistancePriority(for: .vertical), UILayoutPriority(rawValue: UILayoutPriority.RawValue(284)))
         XCTAssertEqual(viewOne.contentHuggingPriority(for: .vertical), UILayoutPriority(rawValue: UILayoutPriority.RawValue(284)))