+++ ConstraidTests/PriorityConstructionTests.swift
@@ -5,7 +5,7 @@ class PriorityConstructionTests: XCTestCase {
     func testConstructsAndSetsPriority() {
         let viewOne = UIView()
         let viewTwo = UIView()
-        let constraint = NSLayoutConstraint(item: viewOne, attribute: .top, relatedBy: .equal, toItem: viewTwo, attribute: .bottom, multiplier: 1.0, constant: 0.0, priority: ConstraidLayoutPriority(rawValue: 22))
+        let constraint = NSLayoutConstraint(item: viewOne, attribute: .top, relatedBy: .equal, toItem: viewTwo, attribute: .bottom, multiplier: 1.0, constant: 0.0, priority: Constraid.LayoutPriority(rawValue: 22))
         XCTAssertEqual(constraint.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraint.firstAttribute, NSLayoutAttribute.top)
         XCTAssertEqual(constraint.relation, NSLayoutRelation.equal)